Fix DeleteCustomer functionality

+++ src/main/java/command/DeleteCustomer.java
@@ -1,6 +1,7 @@
 package command;
 
 import services.CustomerService;
+import utils.CheckCustomers;
 import view.View;
 
 import java.sql.SQLException;
@@ -9,6 +10,7 @@ public class DeleteCustomer implements Command {
     public static final String DELETE_CUSTOMER = "delete customer";
     private final View view;
     CustomerService customerService = new CustomerService();
+    CheckCustomers checkCustomers = new CheckCustomers();
 
     public DeleteCustomer(View view) {
         this.view = view;
@@ -22,18 +24,22 @@ public class DeleteCustomer implements Command {
     @Override
     public void execute() {
         int id;
-        while (true) {
-            try {
-                view.write("Please, enter customer id to delete: ");
-                id = Integer.parseInt(view.read());
-                break;
-            } catch (NumberFormatException e) {
-                view.write("Invalid value. Use digits");
-            }
-        }
         try {
-            customerService.deleteCustomer(id);
-            view.write("Customer with id " + id + " successfully deleted");
+            while (true) {
+                try {
+                    view.write("Please, enter customer id to delete: ");
+                    id = Integer.parseInt(view.read());
+                    if (checkCustomers.IsCustomerIdExists(id)) {
+                        customerService.deleteCustomer(id);
+                        view.write("Customer with id " + id + " successfully deleted");
+                        break;
+                    } else {
+                        System.out.println("Customer id doesn't exists");
+                    }
+                } catch (NumberFormatException e) {
+                    view.write("Invalid value. Use digits");
+                }
+            }
         } catch (SQLException e) {
             e.getStackTrace();
         }
